refactor: pass add/del function props to buttons
To make them more reusable between components.
