+++ includes/publisher/js/src/components/RichText/DeleteItemButton.jsx
@@ -4,14 +4,7 @@ import TrashIcon from "../../../../../components/icons/TrashIcon";
 import { colors } from "../../../../../shared-assets/js/emotion";
 import { __ } from "@wordpress/i18n";
 
-const DeleteItemButton = ({ index, setValues }) => {
-	const deleteItem = () =>
-		setValues((currentValues) => {
-			const newValues = [...currentValues];
-			newValues.splice(index, 1);
-			return newValues;
-		});
-
+const DeleteItemButton = ({ deleteItem }) => {
 	return (
 		<button
 			aria-label={__("Remove item.", "atlas-content-modeler")}
